Move document metadata to metadata.yaml
This nicely separates code and data and will make it easier to keep the
logic of conf.py in a centralized repo while making it more clear what
metadata needs to be maintained for each document.

Also add a requirements.txt to install PyYAML and the LSST design doc
sphinx theme.

For DM-3814.